requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
gspread>=5.10.0
google-auth>=2.22.0
python-dotenv>=1.0.0
loguru>=0.7.0
keyring>=24.0.0
lxml>=4.9.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
